+++ Runtime/IImGuiVisibilityAttribute.cs
@@ -1,0 +1,6 @@
+﻿namespace SaintsField
+{
+    public interface IImGuiVisibilityAttribute
+    {
+    }
+}