+++ Samples~/Scripts/IssueAndTesting/Issue/Issue4/Editor/ShowIfPlayAttributeDrawer.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using System.Reflection;
-using SaintsField.Editor.Drawers;
+using SaintsField.Editor.Drawers.VisibilityDrawers;
 using UnityEditor;
 
 namespace SaintsField.Samples.Scripts.IssueAndTesting.Issue.Issue4.Editor
@@ -8,7 +8,8 @@ namespace SaintsField.Samples.Scripts.IssueAndTesting.Issue.Issue4.Editor
     [CustomPropertyDrawer(typeof(ShowIfPlayAttribute))]
     public class ShowIfPlayAttributeDrawer: VisibilityAttributeDrawer
     {
-        protected override (string error, bool shown) IsShown(SerializedProperty property, VisibilityAttribute visibilityAttribute, FieldInfo info, Type type, object target)
+        protected override (string error, bool shown) IsShown(SerializedProperty property, ISaintsAttribute visibilityAttribute, FieldInfo info,
+            Type type, object target)
         {
             return ("", EditorApplication.isPlaying);
         }