+++ Samples~/Scripts/IssueAndTesting/NA/Issue377.cs
@@ -1,0 +1,61 @@
+using System;
+using System.Collections.Generic;
+using DG.DemiEditor;
+using UnityEngine;
+
+namespace SaintsField.Samples.Scripts.IssueAndTesting.NA
+{
+    public class Issue377 : MonoBehaviour
+    {
+        [Serializable]
+        public class VideoItem
+        {
+            // DeEditorUtils.List titles = new();
+            [Dropdown("titles"), BelowRichLabel(nameof(value), true)]
+            public string value;
+
+            public DropdownList<string> titles = new DropdownList<string>
+            {
+                { "Example1", "Example1" },
+                { "Example2", "Example2" },
+                { "Example3", "Example3" },
+            };
+
+            public VideoItem() {
+                // titles = TimelineCreator.sectionTitles;
+            }
+        }
+
+        // [SaintsRow] public VideoItem[] videoItem;
+
+        [Serializable]
+        public class TimelineItem
+        {
+
+            string name = "";
+            public TimelineItem()
+            {
+
+                name = "TImeline Elemenet";
+            }
+            public bool label;
+            [ShowIf("label")]
+            public string LabelText = "Round 1";
+            [SaintsRow]
+            public List<VideoItem> videos = new List<VideoItem>();
+            public bool loop;
+            [ShowIf("ShouldShowLoop")]
+            public int loopCount=1;
+            public bool startTimer;
+            [ShowIf("startTimer")]
+            public float timerTime = 60;
+            bool ShouldShowLoop()
+            {
+                return loop&&!label;
+            }
+
+        }
+
+        [SerializeField, SaintsRow] List<TimelineItem> Timeline = new();
+    }
+}