+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.1.11
 
 1.  Fix [Issue 3](https://github.com/TylerTemp/SaintsField/issues/3) Texture2D can not be destroyed on a GC calling. 
+2.  UI Toolkit: Fix [Issue 2](https://github.com/TylerTemp/SaintsField/issues/2) incorrect readonly.
 
 ## 2.1.10